requirements.txt
pyyaml
google-cloud-aiplatform
google-auth
kfp
pandas
pyarrow